+++ src/Kiota.Builder/Writers/CSharp/CodeMethodWriter.cs
@@ -259,6 +259,12 @@ public class CodeMethodWriter : BaseElementWriter<CodeMethod, CSharpConventionSe
             {
                 defaultValue = defaultValue.TrimQuotes();
             }
+            else if (defaultValue.StartsWith('"') && defaultValue.EndsWith('"'))
+            {
+                defaultValue = defaultValue[1..^1].Replace("\"", "\\\"", StringComparison.Ordinal).Replace("\n", "\\n", StringComparison.Ordinal);
+                defaultValue = $"\"{defaultValue}\"";
+            }
+
             writer.WriteLine($"{propWithDefault.Name.ToFirstCharacterUpperCase()} = {defaultValue};");
         }
         if (parentClass.IsOfKind(CodeClassKind.RequestBuilder) &&